Move files all under same package
